Start implementing add performance with artist

+++ lib/sing_for_needs/performances/performance.ex
@@ -9,7 +9,7 @@ defmodule SingForNeeds.Performances.Performance do
   schema "performances" do
     field :name, :string
     field :detail, :string
-    field :amount_raised, :float
+    field :amount_raised, :decimal
     many_to_many(:artists, Artist , join_through: :artists_performances, on_replace: :delete)
 
     timestamps()
